pyproject.toml
[tool.poetry]
name = "yat-graphql"
version = "0.1.0"
description = "Yat.Money GraphQL API"
authors = ["YZoZ <mail@yzoz.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.1"
strawberry-graphql = "^0.126.0"
motor = "^3.0.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
pydantic = "^1.9.2"
PyNaCl = "^1.5.0"
aio-pika = "^8.1.1"
pika = "^1.3.0"
pymongo = "^4.2.0"
base58 = "^2.1.1"
redis = {extras = ["hiredis"], version = "^4.4.2"}
aioredis = {extras = ["hiredis"], version = "^2.0.1"}

[tool.poetry.dev-dependencies]
watchfiles = "^0.16.1"
gql = {extras = ["all"], version = "^3.4.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
